Created class ClientSock
I added a class cLinetScok that will be used to connect to the data base.
It uses the base socket api from c++. I made this class to make future used of it a lot cleaner.

I also added some code to the main method that just loads all the preloadData.txt
into the DataBase. This will serve as the data that will be used for the entire program